requirements.txt
PyQt5>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
pyinstaller>=5.13.0
